boringlaunch-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
selenium==4.18.1
webdriver-manager==4.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
loguru==0.7.2
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.1
openpyxl==3.1.2
aiohttp==3.9.3
python-magic==0.4.27
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
pytest-cov==4.1.0
pytest-mock==3.12.0 